# Support RankingComparator and break ranking ties in case of having different attack stats