--- a/pvp_core.go
+++ b/pvp_core.go
@@ -21,7 +21,7 @@ func calculateCpMultiplier(level float64) float64 {
 }
 
 // calculateHp is used to calculate Pokemon HP.
-func calculateHp(stats PokemonStats, stamina int, level float64) int {
+func calculateHp(stats *PokemonStats, stamina int, level float64) int {
 	hp := int(float64(stats.Stamina+stamina) * calculateCpMultiplier(level))
 	if hp <= 10 {
 		return 10
@@ -29,18 +29,8 @@ func calculateHp(stats PokemonStats, stamina int, level float64) int {
 	return hp
 }
 
-// calculateStatProduct is used to calculate Pokemon stat product.
-func calculateStatProduct(stats PokemonStats, attack int, defense int, stamina int, level float64) float64 {
-	multiplier := calculateCpMultiplier(level)
-	hp := math.Floor(float64(stamina+stats.Stamina) * multiplier)
-	if hp < 10 {
-		hp = 10.0
-	}
-	return float64(attack+stats.Attack) * multiplier * float64(defense+stats.Defense) * multiplier * hp
-}
-
-// calculateStatProduct is used to calculate CP for provided Pokemon data.
-func calculateCp(stats PokemonStats, attack int, defense int, stamina int, level float64) int {
+// calculateCp is used to calculate CP for provided Pokemon data.
+func calculateCp(stats *PokemonStats, attack, defense, stamina int, level float64) int {
 	multiplier := calculateCpMultiplier(level)
 
 	cp := int(multiplier * multiplier * float64(stats.Attack+attack) *
@@ -52,11 +42,11 @@ func calculateCp(stats PokemonStats, attack int, defense int, stamina int, level
 }
 
 // calculatePvPStat is core method used to calculate PvP stats for provided Pokemon data.
-func calculatePvPStat(stats PokemonStats, attack int, defense int, stamina int, cap int, lvCap float64, minLevel float64) (Ranking, error) {
+func calculatePvPStat(out *PvPRankingStats, stats *PokemonStats, attack, defense, stamina, cap int, lvCap, minLevel float64) error {
 	bestCP := calculateCp(stats, attack, defense, stamina, minLevel)
 
 	if bestCP > cap {
-		return Ranking{}, ErrPvpStatBestCp
+		return ErrPvpStatBestCp
 	}
 	lowest, highest := minLevel, lvCap
 	for mid := math.Ceil(lowest+highest) / 2; lowest < highest; mid = math.Ceil(lowest+highest) / 2 {
@@ -69,20 +59,23 @@ func calculatePvPStat(stats PokemonStats, attack int, defense int, stamina int, 
 		}
 	}
 
-	return Ranking{
-		Value:   calculateStatProduct(stats, attack, defense, stamina, lowest),
-		Attack:  attack,
-		Defense: defense,
-		Stamina: stamina,
-		Level:   lowest,
-		Cp:      bestCP,
-	}, nil
+	multiplier := calculateCpMultiplier(lowest)
+	out.Attack = float64(attack+stats.Attack) * multiplier
+	hp := math.Floor(float64(stamina+stats.Stamina) * multiplier)
+	if hp < 10 {
+		hp = 10.0
+	}
+	out.Value = out.Attack * float64(defense+stats.Defense) * multiplier * hp
+	out.Level = lowest
+	out.Cp = bestCP
+	return nil
 }
 
 // calculateRanks is core method used to calculate PvP ranks for provided Pokemon data.
-func calculateRanks(stats PokemonStats, cpCap int, lvCap float64) ([16][16][16]Ranking, [4096]Ranking) {
-	var combinations [16][16][16]Ranking
-	var sortedRanks [4096]Ranking
+/*
+func calculateRanks(stats *PokemonStats, cpCap int, lvCap float64, comparator RankingComparator) (*[16][16][16]Ranking, *[4096]*Ranking) {
+	combinations := new([16][16][16]Ranking)
+	sortedRanks := new([4096]*Ranking)
 	var c uint16
 
 	for a := 0; a <= 15; a++ {
@@ -93,56 +86,110 @@ func calculateRanks(stats PokemonStats, cpCap int, lvCap float64) ([16][16][16]R
 					continue
 				}
 				combinations[a][d][s] = currentStat
-				sortedRanks[c] = currentStat
+				sortedRanks[c] = &currentStat
 				c++
 			}
 		}
 	}
 
-	sort.Sort(rankingSortable(sortedRanks[:]))
+	sort.SliceStable(sortedRanks, func(i, j int) bool {
+		return comparator(sortedRanks[i], sortedRanks[j]) < 0
+	})
 
 	best := sortedRanks[0].Value
 	var i, j int16
 	for i, j = 0, 0; i < int16(len(sortedRanks)); i++ {
-		entry := &sortedRanks[i]
-		combinationsEntry := &combinations[entry.Attack][entry.Defense][entry.Stamina]
-
+		entry := sortedRanks[i]
 		percentage := roundFloat(entry.Value/best, 5)
-
 		entry.Percentage = percentage
-		combinationsEntry.Percentage = percentage
-
 		if entry.Value < sortedRanks[j].Value {
 			j = i
 		}
 		rank := j + 1
-
 		entry.Rank = rank
-		combinationsEntry.Rank = rank
 	}
 	return combinations, sortedRanks
 }
+*/
+
+func RankingComparator_Default(a, b *PvPRankingStats) int {
+	d := b.Value - a.Value
+	if d > 0 {
+		return 1
+	}
+	if d < 0 {
+		return -1
+	}
+	d = b.Attack - a.Attack
+	if d > 0 {
+		return 1
+	}
+	if d < 0 {
+		return -1
+	}
+	return 0
+}
+func RankingComparator_PreferHigherCp(a, b *PvPRankingStats) int {
+	d := RankingComparator_Default(a, b)
+	if d > 0 {
+		return 1
+	}
+	if d < 0 {
+		return -1
+	}
+	d = b.Cp - a.Cp
+	if d > 0 {
+		return 1
+	}
+	if d < 0 {
+		return -1
+	}
+	return 0
+}
+func RankingComparator_PreferLowerCp(a, b *PvPRankingStats) int {
+	d := RankingComparator_Default(a, b)
+	if d > 0 {
+		return 1
+	}
+	if d < 0 {
+		return -1
+	}
+	d = a.Cp - b.Cp
+	if d > 0 {
+		return 1
+	}
+	if d < 0 {
+		return -1
+	}
+	return 0
+}
 
 // calculateRanksCompact is optimized (for cache) core method used to calculate PvP ranks for provided Pokemon data.
-func calculateRanksCompact(stats PokemonStats, cpCap int, lvCap float64, ivFloor int) ([4096]int16, [4096]Ranking) {
-	var combinations [4096]int16
-	var sortedRanks [4096]Ranking
+func calculateRanksCompact(stats *PokemonStats, cpCap int, lvCap float64, comparator RankingComparator, ivFloor int) (*[4096]int16, *[4096]PvPRankingStats) {
+	combinations := new([4096]int16)
+	sortedRanks := new([4096]PvPRankingStats)
 
+	count := 0
 	for a := ivFloor; a <= 15; a++ {
 		for d := ivFloor; d <= 15; d++ {
 			for s := ivFloor; s <= 15; s++ {
-				entry, _ := calculatePvPStat(stats, a, d, s, cpCap, lvCap, 1)
-				entry.Index = (a*16+d)*16 + s
-				sortedRanks[entry.Index] = entry
+				if calculatePvPStat(&sortedRanks[count], stats, a, d, s, cpCap, lvCap, 1) == nil {
+					sortedRanks[count].Index = (a*16+d)*16 + s
+					count++
+				}
 			}
 		}
 	}
 
-	sort.Sort(rankingSortableIndexed(sortedRanks[:]))
+	// use Slice over SliceStable for performance since we have index available to us
+	sort.Slice(sortedRanks[:count], func(i, j int) bool {
+		d := comparator(&sortedRanks[i], &sortedRanks[j])
+		return d < 0 || d == 0 && sortedRanks[i].Index < sortedRanks[j].Index
+	})
 
 	for i, j := 0, 0; i < len(sortedRanks); i++ {
 		entry := &sortedRanks[i]
-		if entry.Value < sortedRanks[j].Value {
+		if comparator(&sortedRanks[j], entry) < 0 {
 			j = i
 		}
 		combinations[entry.Index] = int16(j + 1)