Many bug fixes included. Commented out non-functional `CalculateAllRanks` and `CalculateTopRanks`.

Removed nonsensical test cases like testing level 0 stuff.

Fixes https://github.com/UnownHash/gohbem/issues/9. Mirrors https://github.com/Mygod/ohbem/pull/16.