@Mygod 2023-04-05
Before changing to `sort.Sort`:
```
BenchmarkQueryPvPRank-16                   22726           3200570 ns/op
BenchmarkQueryPvPRankCached-16          64416926              1072 ns/op
```

After:
```
BenchmarkQueryPvPRank-16                   26271           2741242 ns/op
BenchmarkQueryPvPRankCached-16          61786150              1050 ns/op
```

Master branch:
```
BenchmarkQueryPvPRank-16                   24817           2860125 ns/op
BenchmarkQueryPvPRankCached-16          43727300              1685 ns/op
```